requirement.txt
asgiref==3.6.0
boto3==1.26.98
botocore==1.29.98
Brotli==1.0.9
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
ConfigArgParse==1.5.3
cx-Oracle==8.3.0
Django==3.2.18
django-environ==0.10.0
django-filter==23.1
django-storages==1.13.2
Flask==2.2.3
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
gevent==22.10.2
geventhttpclient==2.0.9
greenlet==2.0.2
idna==3.4
importlib-metadata==6.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
locust==2.15.1
MarkupSafe==2.1.2
msgpack==1.0.5
mssql-django==1.2
mysqlclient==2.1.1
Pillow==9.5.0
psutil==5.9.4
pyodbc==4.0.39
python-dateutil==2.8.2
pytz==2022.7.1
pyzmq==25.0.2
requests==2.28.2
roundrobin==0.0.4
s3transfer==0.6.0
six==1.16.0
sorl-thumbnail==12.9.0
sqlparse==0.4.3
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
zipp==3.15.0
zope.event==4.6
zope.interface==6.0
